requirements.txt
colorama
mysql-connector-python
keyboard